[Android] Exclude passive provider
The passive provider is always active (which confuses IsLocationEnabled), is
spotty for when it works and is a specialized Android feature. This doesn't
really meet the design goals of Geolocator and for now will be excluded.

Also, convert providers container to a .NET array, should improve
performance in some spots in Geolocator.