[WP] Do not trigger geolocation in ctor
Not only does this value not even get used, it will trigger geolocation in
the ctor, which is bad for performance and makes it harder to cope with
Windows Phone guidelines that you, the developer, must provide the user a
way to disable geolocation. It's easier to not call a method than it is to
not initialize an object.