Add MediaFileNotFoundException
Why? Because Android.
Android's intent system is quite loose. There are no safeguards around
what an the pick and capture intents will return. For example, picking a
Picasa image out of the Android gallery app will return a content URI that
does not store its file location in the DATA column like other images do.
Other apps are free to do whatever they want.

We can't possible handle every edge case for every camera or gallery app.
So, we'll try to handle the basics, and for everything else we will return
this exception which effectively says "the user completed the action,
here's the URI, we don't know what to do with it.. you try." It's better
than random crashes inside Xamarin.Mobile. We may still get some, but
we'll try to filter everything we don't know what to do with to this
exception.

The reason we're not just handing back the uri through the normal means
is the user excepts MediaFile.Path to be a normal file path and should
never need special handling on a platform basis.