Asyncify android MediaPicker internals
The ContentResolver content stream could come from anywhere, we should
copy that on a background thread and take advantage of the fact that
everything is Task<T>'d.

Also, finish the activity immediately. This should avoid any issues with
our activity still being open when the completion fires. Haven't seen this
happen in practice, but it's a possibility.

We also start invoking the dispose functionality of MediaFile in the cases
where we copy the file. Fixes dangling files from previous copy commit.