Don't dispose the video for now
Unfortunately this sample has been demonstrating an incorrect usage for
cleaning up files for a while. When we redo the Android sample, we should
separate the movie player as the iOS sample is so we have a definitive
place to cleanup.

Going forward the API and docs are consistent on the behavior now so
hopefully this won't bite too many people.