[AD] Implement new MediaPicker API for #2
A lot of the async retrieval logic has been staticized for use outside
MediaPickerActivity. Not the cleanest, but it's all internal we can clean
it up later.

In order to be able to [Obsolete] the old APIs instead of just removing
them we needed to maintain and split the existing code. Things can be
simplified quite a bit once we remove it completely.

There were also some style changes along the way to bring it into line
with Mono standards before we open source.