[iOS] Add Get*UI MediaPicker methods for #2
I haven't [Obsolete]d the old methods here. Although they are not a best
practice, they do work and are simpler. Furthermore Windows Phone and
WinRT currently won't be getting an equivalent API to the new one.

As with the Android changes, there are some mixed in style changes in the
code that I touched to bring it in to the Mono style.